feat: merge the `sender` and `severity` in the DDS payload into the same field to simplify the protocol (#871)

+++ yazi-plugin/src/pubsub/pubsub.rs
@@ -27,7 +27,7 @@ impl Pubsub {
 
 		ps.raw_set(
 			"pub_static",
-			lua.create_function(|_, (severity, kind, value): (u8, mlua::String, Value)| {
+			lua.create_function(|_, (severity, kind, value): (u16, mlua::String, Value)| {
 				if severity < 1 {
 					return Err("Severity must be at least 1").into_lua_err();
 				}
